% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_tss.R
\name{deep_tss}
\alias{deep_tss}
\title{deepTSSscrubbeR constructor function}
\usage{
deep_tss(bam)
}
\arguments{
\item{bam}{Bam file from five-prime mapping data}
}
\description{
deepTSSscrubbeR constructor function
}
