% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{encode_genomic}
\alias{encode_genomic}
\title{One Hot Encode Genomic Sequences}
\usage{
encode_genomic(deep_obj)
}
\arguments{
\item{deep_obj}{deep_tss object}
}
\value{
one hot encoded tensor of sequence stored in @encoded_sequences
}
\description{
One hot encode the genomic sequences from the expanded ranges
}
