% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_sequences.R
\name{encode_sequences}
\alias{encode_sequences}
\title{One Hot Encode Sequences}
\usage{
encode_sequences(deep_obj)
}
\arguments{
\item{experiment}{deep_tss object}
}
\value{
one hot encoded tensor of sequence stored in @encoded_sequences
}
\description{
One hot encode the sequences from the expanded ranges
}
