% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{encode_soft}
\alias{encode_soft}
\title{One Hoe Encode Soft-clipped Bases}
\usage{
encode_soft(deep_obj)
}
\arguments{
\item{deep_obj}{object}
}
\description{
One hot encode the soft-clipped sequences from the bam file
}
