% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_status.R
\name{encode_status}
\alias{encode_status}
\title{Encode TSS Status}
\usage{
encode_status(experiment)
}
\arguments{
\item{experiment}{deep_tss object}
}
\value{
status encoded into tensor in slot @status_encoded
}
\description{
Encode TSS status into proper tensor
}
