% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_ranges.R
\name{expand_range}
\alias{expand_range}
\title{Expand Ranges Function}
\usage{
expand_range(grange, expand_size)
}
\arguments{
\item{grange}{GRange object to expand}

\item{expand_size}{Size to expand ranges on each side of TSS}
}
\value{
Expanded GRanges object
}
\description{
Backend function to expand granges
}
