% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\name{export_raw}
\alias{export_raw}
\title{Export Raw}
\usage{
export_raw(deep_obj, sequence_file, signal_file)
}
\arguments{
\item{deep_obj}{tss_obj}
}
\description{
Export the raw values that eventially get encoded.
}
