% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_sequences.R
\name{get_sequences}
\alias{get_sequences}
\title{Retrieve Sequences}
\usage{
get_sequences(deep_obj, assembly)
}
\arguments{
\item{deep_obj}{dep_tss object}

\item{assembly}{Path to genome assembly}
}
\description{
Retrieve sequences for expanded ranges
}
