% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_shape.R
\name{get_shape}
\alias{get_shape}
\title{Retrieve DNA Shape}
\usage{
get_shape(deep_obj, assembly)
}
\arguments{
\item{deep_obj}{deep_tss object}

\item{assembly}{Path to genome assembly}
}
\description{
Use DNAshapeR to get shape information on signal surrounding TSSs
}
