% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_status.R
\name{mark_status}
\alias{mark_status}
\title{Mark Status}
\usage{
mark_status(deep_obj, lower = 2, upper = 10, use_annotation = FALSE)
}
\arguments{
\item{deep_obj}{deep_tss object}

\item{lower}{Reads less than or equal to 'lower' will be marked as spurious}

\item{upper}{Reads greater than or equal to 'upper' will be marked as good}

\item{use_annotation}{If TRUE will ignore upper and lower bounds and instead
pick a sample of TSSs within and outside of annotated promoters}
}
\value{
GRanges object with status column indicating spurious (1) or not (0)
}
\description{
Mark status as spurious based on read threshold
}
