% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{one_hot_seqs}
\alias{one_hot_seqs}
\title{One Hot Encode a Sequence}
\usage{
one_hot_seqs(sequence)
}
\arguments{
\item{sequence}{String representation of bases}
}
\value{
one hot encoded sequence
}
\description{
Internal function to one hot encode a sequence
}
