% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_sequences.R
\name{retrieve_sequences}
\alias{retrieve_sequences}
\title{Retrieve Sequences}
\usage{
retrieve_sequences(experiment, assembly)
}
\arguments{
\item{experiment}{deep_tss object}

\item{assembly}{File path to genome assembly}
}
\value{
Sequences stored in @sequences
}
\description{
Retrieve sequences for expanded GRanges
}
