% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{Train/Test Split}
\usage{
split_data(experiment, train_split, test_split)
}
\arguments{
\item{experiment}{deep_tss object}

\item{train_split}{Total number of TSSs to use on training set, split evenly between marked status}

\item{test_split}{Total number of TSSs to use on test set, split evenly between marked status}
}
\value{
Training and Test sets saved into slot @split
}
\description{
Split data into training and test set
}
