% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{tss_annotate}
\alias{tss_annotate}
\title{Annotate the TSSs}
\usage{
tss_annotate(deep_obj, annotation, upstream = 1000, downstream = 1000)
}
\arguments{
\item{deep_obj}{deep tss object}

\item{annotation}{genome annotation}

\item{upstream}{Bases upstream to consider promoter}

\item{downstream}{Bases downstream to consider promoter}
}
\description{
Annotate TSSs to the nearest transcript.
}
