% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_model.R
\name{tss_evaluate}
\alias{tss_evaluate}
\title{Evaluate TSS Model}
\usage{
tss_evaluate(deep_obj)
}
\arguments{
\item{deep_obj}{tss_obj with trained model}
}
\description{
Evaluate the trained deepTSSScrubbeR model
}
