% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_model.R
\name{tss_predict}
\alias{tss_predict}
\title{Predict TSS artifact probabilities}
\usage{
tss_predict(deep_obj)
}
\arguments{
\item{deep_obj}{deep_tss object with trained model}
}
\description{
Predict the probability of being an artifact based on the deepTSSscrubbeR model
}
