% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_model.R
\name{tss_train}
\alias{tss_train}
\title{Train TSS Model}
\usage{
tss_train(deep_obj, epochs = 25, batch_size = 150,
  validation_split = 0.25)
}
\arguments{
\item{deep_obj}{tss_obj}

\item{epochs}{epochs}

\item{batch_size}{batch size}

\item{validation_split}{validation split}
}
\description{
Train the deepTSSscrubbeR model
}
