% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_shape.R
\name{encode_shape}
\alias{encode_shape}
\title{Retrieve DNA Shape}
\usage{
encode_shape(deep_obj)
}
\arguments{
\item{deep_obj}{deep_tss object}
}
\description{
Use DNAshapeR to get shape information on signal surrounding TSSs
}
