% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_status.R
\name{mark_status}
\alias{mark_status}
\title{Mark Status}
\usage{
mark_status(experiment, threshold = 1)
}
\arguments{
\item{experiment}{deep_tss object}

\item{threshold}{Reads less than or equal to threshold will be marked as spurious}
}
\value{
GRanges object with status column indicating spurious (1) or not (0)
}
\description{
Mark status as spurious based on read threshold
}
