% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{tss_model}
\alias{tss_model}
\title{Build TSS Model}
\usage{
tss_model(deep_obj, optimizer = "adam", metric = "accuracy")
}
\arguments{
\item{deep_obj}{depp_tss object}

\item{optimizer}{Optimizer for compiling model}

\item{metric}{Metric to optimize model}
}
\description{
Build model of TSS Confidence
}
