% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_model.R
\name{tss_model}
\alias{tss_model}
\title{Build TSS Model}
\usage{
tss_model(deep_obj, optimizer = "adam")
}
\arguments{
\item{deep_obj}{deep_tss object}

\item{optimizer}{Optimizer for compiling model}
}
\description{
Build model of TSS Confidence
}
